WIP: Simplify sagas into a single file for now
