+++ packages/components/src/redux/sagas.js
@@ -1,12 +1,27 @@
-import { takeEvery, call, put } from "redux-saga/effects";
-import * as Api from "./api";
-import * as mocks from "../../mocks";
-import Player from "../../libs/audio-player/";
-const AudioPlayer = new Player();
+import axios from "axios";
+import { call, put, takeEvery } from "redux-saga/effects";
+import * as mocks from "../mocks";
+
+
+const base = process.env.REACT_APP_API_URL || "https://cast-bucket-api.now.sh";
+const mountpoint = process.env.REACT_APP_API_VERSION || "v1";
+
+const api = `${base}/${mountpoint}`;
+
+export function* fetch(url) {
+  try {
+    const { data } = yield call(axios.get, `${api}${url}`);
+    yield put({ type: "FETCH_SUCCEEDED", data });
+    return data;
+  } catch (error) {
+    yield put({ type: "FETCH_FAILED", error });
+  }
+}
+
 
 export function* fetchCategories() {
   try {
-    const categories = yield call(Api.fetch, "/categories");
+    const categories = yield call(fetch, "/categories");
     yield put({ type: "RECEIVED_CATEGORIES", categories });
   } catch (error) {
     yield put({ type: "FETCH_CATEGORIES_FAILED", error });
@@ -20,7 +35,7 @@ export function* fetchCategories() {
 export function* fetchPodcasts(podcastType) {
   console.log("fetching podcasts");
   try {
-    // const podcasts = yield call(Api.fetch, "/podcasts");
+    // const podcasts = yield call(fetch, "/podcasts");
     const podcasts = mocks.podcastListItems.map(item => ({ ...item, type: "PODCAST_ITEM" }));
     yield put({ type: "RECEIVED_PODCASTS", podcasts });
   } catch (error) {
@@ -40,7 +55,8 @@ export function* fetchEpisodes(podcastId) {
 export function* playEpisode({ episodeId, meta }) {
   try {
     const episodeIndex = meta && meta.index ? meta.index : null;
-    AudioPlayer.play(episodeId, episodeIndex);
+    // AudioPlayer.play(episodeId, episodeIndex);
+
     yield put({ type: "PLAY_EPISODE_SUCCESS", episodeId });
   } catch (error) {
     yield put({ type: "PLAY_EPISODE_FAILED", episodeId, error });
@@ -50,7 +66,7 @@ export function* playEpisode({ episodeId, meta }) {
 export function* pauseEpisode({ episodeId, meta }) {
   try {
     const episodeIndex = meta && meta.index ? meta.index : null;
-    AudioPlayer.pause(episodeId, episodeIndex);
+    // AudioPlayer.pause(episodeId, episodeIndex);
     yield put({ type: "PAUSE_EPISODE_SUCCESS", episodeId });
   } catch (error) {
     yield put({ type: "PAUSE_EPISODE_FAILED", error });
@@ -59,6 +75,15 @@ export function* pauseEpisode({ episodeId, meta }) {
 
 export function* playEpisodeFailed({ error, episodeId }) {
   console.error(`Unable to play episode ${episodeId} due to ${error.message}`);
+}
+
+export function* getLastPlayedEpisode() {
+  try {
+    const lastPlayedEpisode = AudioPlayer.getLastPlayedEpisode();
+    console.log('>>>-SHRIRAM->>> lastPlayedEpisode', lastPlayedEpisode);
+  } catch (error) {
+    
+  }
 }
 
 export default function* rootSaga() {
